requirements.txt
asgiref==3.6.0
Django==4.1.5
django-crispy-forms==1.14.0
django-dotenv==1.4.2
django-froala-editor==4.0.17
Pillow==9.4.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
